Move demos to r3f folder
