+++ .eslintrc.json
@@ -31,7 +31,7 @@
 		"no-constant-condition": "off",
 		// "no-constant-condition": [ "error", { "checkLoops": "exceptWhileTrue" } ],
 
-		"no-unused-vars": [ 1, { "args": "none" } ],
+		"no-unused-vars": [ "error", { "args": "none" } ],
 		"no-throw-literal": [ "error" ],
 		"prefer-const": [ "error", { "destructuring": "any", "ignoreReadBeforeAssign": false } ],
 		"quotes": [ "error", "single" ]
